Notificando clinica dos clientes que foram avisados. Altera tipo de dado da hora para comparar corretamente. Inclui na mensagem a descricao de que eh msg automatica e nao precisa de resposta.

+++ app/Http/Controllers/Agenda.php
@@ -24,7 +24,7 @@ class Agenda extends Controller
 
         try {
             // Ex.: Se a data é 2022-09-08 10:40:00, resultado sera 10
-            $hora_atual = explode(":", explode(" ", date("Y-m-d H:i:s"))[1])[0];
+            $hora_atual = (int)explode(":", explode(" ", date("Y-m-d H:i:s"))[1])[0];
 
             if(\in_array($hora_atual, $horario_min)){
                 die("Nao notificar cliente às ". $hora_atual . "h. Muito cedo!");
